+++ JumpScale9/clients/git/GitFactory.py
@@ -7,11 +7,365 @@ class GitFactory:
 
     def __init__(self):
         self.logger = j.logger.get('j.clients.git')
-        self.pullGitRepo = j.do.pullGitRepo
-        self.getGitRepoArgs = j.do.getGitRepoArgs
-        self.rewriteGitRepoUrl = j.do.rewriteGitRepoUrl
-        self.getGitBranch = j.do.getGitBranch
         self.__jslocation__ = "j.clients.git"
+
+    def rewriteGitRepoUrl(self, url="", login=None, passwd=None, ssh="auto"):
+        """
+        Rewrite the url of a git repo with login and passwd if specified
+
+        Args:
+            url (str): the HTTP URL of the Git repository. ex: 'https://github.com/odoo/odoo'
+            login (str): authentication login name
+            passwd (str): authentication login password
+            ssh = if True will build ssh url, if "auto" will check if there is ssh-agent available & keys are loaded,
+                if yes will use ssh
+
+        Returns:
+            (repository_host, repository_type, repository_account, repository_name, repository_url)
+        """
+
+        if ssh == "auto" or ssh == "first":
+            ssh = self.SSHAgentAvailable()
+        elif ssh or ssh is False:
+            pass
+        else:
+            raise RuntimeError(
+                "ssh needs to be auto, first or True or False: here:'%s'" %
+                ssh)
+
+        url_pattern_ssh = re.compile('^(git@)(.*?):(.*?)/(.*?)/?$')
+        sshmatch = url_pattern_ssh.match(url)
+        url_pattern_http = re.compile('^(https?://)(.*?)/(.*?)/(.*?)/?$')
+        httpmatch = url_pattern_http.match(url)
+        if not sshmatch:
+            match = httpmatch
+        else:
+            match = sshmatch
+
+        if not match:
+            raise RuntimeError(
+                "Url is invalid. Must be in the form of 'http(s)://hostname/account/repo' or 'git@hostname:account/repo'")
+
+        protocol, repository_host, repository_account, repository_name = match.groups()
+        if protocol.startswith("git") and ssh is False:
+            protocol = "https://"
+
+        if not repository_name.endswith('.git'):
+            repository_name += '.git'
+
+        if login == 'ssh' or ssh:
+            repository_url = 'git@%(host)s:%(account)s/%(name)s' % {
+                'host': repository_host,
+                'account': repository_account,
+                'name': repository_name,
+            }
+            protocol = "ssh"
+
+        elif login and login != 'guest':
+            repository_url = '%(protocol)s%(login)s:%(password)s@%(host)s/%(account)s/%(repo)s' % {
+                'protocol': protocol,
+                'login': login,
+                'password': passwd,
+                'host': repository_host,
+                'account': repository_account,
+                'repo': repository_name,
+            }
+
+        else:
+            repository_url = '%(protocol)s%(host)s/%(account)s/%(repo)s' % {
+                'protocol': protocol,
+                'host': repository_host,
+                'account': repository_account,
+                'repo': repository_name,
+            }
+        if repository_name.endswith(".git"):
+            repository_name = repository_name[:-4]
+
+        return protocol, repository_host, repository_account, repository_name, repository_url
+
+    def getGitRepoArgs(
+            self,
+            url="",
+            dest=None,
+            login=None,
+            passwd=None,
+            reset=False,
+            branch=None,
+            ssh="auto",
+            codeDir=None,
+            executor=None):
+        """
+        Extracts and returns data useful in cloning a Git repository.
+
+        Args:
+            url (str): the HTTP/GIT URL of the Git repository to clone from. eg: 'https://github.com/odoo/odoo.git'
+            dest (str): the local filesystem path to clone to
+            login (str): authentication login name (only for http)
+            passwd (str): authentication login password (only for http)
+            reset (boolean): if True, any cached clone of the Git repository will be removed
+            branch (str): branch to be used
+            ssh if auto will check if ssh-agent loaded, if True will be forced to use ssh for git
+
+        # Process for finding authentication credentials (NOT IMPLEMENTED YET)
+
+        - first check there is an ssh-agent and there is a key attached to it, if yes then no login & passwd will be used & method will always be git
+        - if not ssh-agent found
+            - then we will check if url is github & ENV argument GITHUBUSER & GITHUBPASSWD is set
+                - if env arguments set, we will use those & ignore login/passwd arguments
+            - we will check if login/passwd specified in URL, if yes willl use those (so they get priority on login/passwd arguments)
+            - we will see if login/passwd specified as arguments, if yes will use those
+        - if we don't know login or passwd yet then
+            - login/passwd will be fetched from local git repo directory (if it exists and reset==False)
+        - if at this point still no login/passwd then we will try to build url with anonymous
+
+
+        # Process for defining branch
+
+        - if branch arg: None
+            - check if git directory exists if yes take that branch
+            - default to 'master'
+        - if it exists, use the branch arg
+
+        Returns:
+            (repository_host, repository_type, repository_account, repository_name, dest, repository_url)
+
+            - repository_type http or git
+
+        Remark:
+            url can be empty, then the git params will be fetched out of the git configuration at that path
+        """
+
+        if url == "":
+            if dest is None:
+                raise RuntimeError("dest cannot be None (url is also '')")
+            if not self.exists(dest):
+                raise RuntimeError(
+                    "Could not find git repo path:%s, url was not specified so git destination needs to be specified." %
+                    (dest))
+
+        if login is None and url.find("github.com/") != -1:
+            # can see if there if login & passwd in OS env
+            # if yes fill it in
+            if "GITHUBUSER" in os.environ:
+                login = os.environ["GITHUBUSER"]
+            if "GITHUBPASSWD" in os.environ:
+                passwd = os.environ["GITHUBPASSWD"]
+
+        protocol, repository_host, repository_account, repository_name, repository_url = self.rewriteGitRepoUrl(
+            url=url, login=login, passwd=passwd, ssh=ssh)
+
+        repository_type = repository_host.split(
+            '.')[0] if '.' in repository_host else repository_host
+
+        if not dest:
+            if codeDir is None:
+                if not executor:
+                    codeDir = j.CODEDIR
+                else:
+                    codeDir = executor.prefab.core.dir_paths['CODEDIR']
+            dest = '%(codedir)s/%(type)s/%(account)s/%(repo_name)s' % {
+                'codedir': codeDir,
+                'type': repository_type.lower(),
+                'account': repository_account.lower(),
+                'repo_name': repository_name,
+            }
+
+        if reset:
+            self.delete(dest)
+
+        # self.createDir(dest)
+
+        return repository_host, repository_type, repository_account, repository_name, dest, repository_url
+
+    def pullGitRepo(
+            self,
+            url="",
+            dest=None,
+            login=None,
+            passwd=None,
+            depth=None,
+            ignorelocalchanges=False,
+            reset=False,
+            branch=None,
+            tag=None,
+            revision=None,
+            ssh="auto",
+            executor=None,
+            codeDir=None,
+            timeout=600):
+        """
+        will clone or update repo
+        if dest is None then clone underneath: /opt/code/$type/$account/$repo
+        will ignore changes !!!!!!!!!!!
+
+        @param ssh ==True means will checkout ssh
+        @param ssh =="first" means will checkout sss first if that does not work will go to http
+        """
+        if branch == "":
+            branch = None
+        if branch is not None and tag is not None:
+            raise RuntimeError("only branch or tag can be set")
+
+        if ssh == "first" or ssh == "auto":
+            try:
+                return self.pullGitRepo(
+                    url,
+                    dest,
+                    login,
+                    passwd,
+                    depth,
+                    ignorelocalchanges,
+                    reset,
+                    branch,
+                    tag=tag,
+                    revision=revision,
+                    ssh=True,
+                    executor=executor)
+            except Exception as e:
+                base, provider, account, repo, dest, url = self.getGitRepoArgs(
+                    url, dest, login, passwd, reset=reset, ssh=False, codeDir=codeDir, executor=executor)
+                return self.pullGitRepo(
+                    url,
+                    dest,
+                    login,
+                    passwd,
+                    depth,
+                    ignorelocalchanges,
+                    reset,
+                    branch,
+                    tag=tag,
+                    revision=revision,
+                    ssh=False,
+                    executor=executor)
+
+        base, provider, account, repo, dest, url = self.getGitRepoArgs(
+            url, dest, login, passwd, reset=reset, ssh=ssh, codeDir=codeDir, executor=executor)
+
+        self.logger.info("%s:pull:%s ->%s" % (executor, url, dest))
+
+        existsDir = self.exists(
+            dest) if not executor else executor.exists(dest)
+
+        checkdir = "%s/.git" % (dest)
+        existsGit = self.exists(
+            checkdir) if not executor else executor.exists(checkdir)
+
+        if existsDir:
+            if not existsGit:
+                raise RuntimeError("found directory but .git not found in %s" % dest)
+            # if we don't specify the branch, try to find the currently
+            # checkedout branch
+            cmd = 'cd %s; git rev-parse --abbrev-ref HEAD' % dest
+            rc, out, err = self.execute(
+                cmd, die=False, showout=False, executor=executor)
+            if rc == 0:
+                branchFound = out.strip()
+            else:  # if we can't retreive current branch, use master as default
+                branchFound = 'master'
+                # raise RuntimeError("Cannot retrieve branch:\n%s\n" % cmd)
+            if branch is not None and branch != branchFound and ignorelocalchanges is False:
+                raise RuntimeError(
+                    "Cannot pull repo '%s', branch on filesystem is not same as branch asked for.\n"
+                    "Branch asked for: %s\n"
+                    "Branch found: %s\n"
+                    "To choose other branch do e.g:"
+                    "export JSBRANCH='%s'\n" %
+                    (repo, branch, branchFound, branchFound))
+
+            if ignorelocalchanges:
+                self.logger.info(
+                    ("git pull, ignore changes %s -> %s" %
+                     (url, dest)))
+                cmd = "cd %s;git fetch" % dest
+                if depth is not None:
+                    cmd += " --depth %s" % depth
+                    self.execute(cmd, executor=executor)
+                if branch is not None:
+                    self.logger.info("reset branch to:%s" % branch)
+                    self.execute(
+                        "cd %s;git fetch; git reset --hard origin/%s" %
+                        (dest, branch), timeout=timeout, executor=executor)
+
+            else:
+
+                if branch is None and tag is None:
+                    branch = branchFound
+
+                # pull
+                self.logger.info(("git pull %s -> %s" % (url, dest)))
+                if url.find("http") != -1:
+                    cmd = "mkdir -p %s;cd %s;git -c http.sslVerify=false pull origin %s" % (
+                        dest, dest, branch)
+                    self.logger.info(cmd)
+                    self.execute(cmd, timeout=timeout, executor=executor)
+                else:
+                    try:
+                        cmd = "cd %s;git pull origin %s" % (dest, branch)
+                        self.logger.info(cmd)
+                        self.execute(cmd, timeout=timeout, executor=executor)
+                    except Exception as e:
+                        protocol, host, account, repo_name, repo_url = self.rewriteGitRepoUrl(url=url, ssh=False)
+                        cmd = "cd %s;git -c http.sslVerify=false pull %s %s" % (dest, repo_url, branch)
+                        self.logger.info(cmd)
+                        self.execute(cmd, timeout=timeout, executor=executor)
+
+        else:
+            self.logger.info(("git clone %s -> %s" % (url, dest)))
+            # self.createDir(dest)
+            extra = ""
+            if depth is not None:
+                extra = "--depth=%s" % depth
+            if url.find("http") != -1:
+                if branch is not None:
+                    cmd = "mkdir -p %s;cd %s;git -c http.sslVerify=false clone %s -b %s %s %s" % (
+                        self.getParent(dest), self.getParent(dest), extra, branch, url, dest)
+                else:
+                    cmd = "mkdir -p %s;cd %s;git -c http.sslVerify=false clone %s  %s %s" % (
+                        self.getParent(dest), self.getParent(dest), extra, url, dest)
+            else:
+                if branch is not None:
+                    cmd = "mkdir -p %s;cd %s;git clone %s -b %s %s %s" % (
+                        self.getParent(dest), self.getParent(dest), extra, branch, url, dest)
+                else:
+                    cmd = "mkdir -p %s;cd %s;git clone %s  %s %s" % (
+                        self.getParent(dest), self.getParent(dest), extra, url, dest)
+
+            self.logger.info(cmd)
+
+            # self.logger.info(str(executor)+" "+cmd)
+            self.execute(cmd, timeout=timeout, executor=executor)
+
+        if tag is not None:
+            self.logger.info("reset tag to:%s" % tag)
+            self.execute("cd %s;git checkout tags/%s" %
+                         (dest, tag), timeout=timeout, executor=executor)
+
+        if revision is not None:
+            cmd = "mkdir -p %s;cd %s;git checkout %s" % (dest, dest, revision)
+            self.logger.info(cmd)
+            self.execute(cmd, timeout=timeout, executor=executor)
+
+        return dest
+
+    def getGitBranch(self, path):
+
+        # if we don't specify the branch, try to find the currently checkedout
+        # branch
+        cmd = 'cd %s;git rev-parse --abbrev-ref HEAD' % path
+        try:
+            rc, out, err = self.execute(cmd, showout=False, outputStderr=False)
+            if rc == 0:
+                branch = out.strip()
+            else:  # if we can't retreive current branch, use master as default
+                branch = 'master'
+        except BaseException:
+            branch = 'master'
+
+        return branch
+
+
+
+
 
     def parseUrl(self, url):
         """
