+++ JumpScale9/clients/ssh/SSHClientFactory.py
@@ -12,21 +12,6 @@ class SSHClientFactory:
 
     logger = j.logger.get("j.clients.ssh")
 
-    # to not have to duplicate information
-    SSHKeysLoad = j.do.SSHKeysLoad
-    _addSSHAgentToBashProfile = j.do._addSSHAgentToBashProfile
-    _initSSH_ENV = j.do._initSSH_ENV
-    _getSSHSocketpath = j.do._getSSHSocketpath
-    SSHKeysLoad = j.do.SSHKeysLoad
-    SSHKeyGetPathFromAgent = j.do.SSHKeyGetPathFromAgent
-    SSHKeyGetFromAgentPub = j.do.SSHKeyGetFromAgentPub
-    SSHKeysListFromAgent = j.do.SSHKeysListFromAgent
-    SSHEnsureKeyname = j.do.SSHEnsureKeyname
-    authorize_user = j.do.authorize_user
-    authorize_root = j.do.authorize_root
-    SSHAuthorizeKey = j.do.SSHAuthorizeKey
-    _loadSSHAgent = j.do._loadSSHAgent
-    SSHAgentAvailable = j.do.SSHAgentAvailable
 
     def __init__(self):
         self.__jslocation__ = "j.clients.ssh"
@@ -153,3 +138,406 @@ class SSHClientFactory:
         with self._lock:
             for key, client in self.cache.items():
                 client.close()
+
+    def _addSSHAgentToBashProfile(self, path=None):
+
+        bashprofile_path = os.path.expanduser("~/.bash_profile")
+        if not self.exists(bashprofile_path):
+            self.execute('touch %s' % bashprofile_path)
+
+        content = self.readFile(bashprofile_path)
+        out = ""
+        for line in content.split("\n"):
+            if line.find("#JSSSHAGENT") != -1:
+                continue
+            if line.find("SSH_AUTH_SOCK") != -1:
+                continue
+
+            out += "%s\n" % line
+
+        if "SSH_AUTH_SOCK" in os.environ:
+            self.logger.info("NO NEED TO ADD SSH_AUTH_SOCK to env")
+            self.writeFile(bashprofile_path, out)
+            return
+
+        # out += "\njs 'j.do._.loadSSHAgent()' #JSSSHAGENT\n"
+        out += "export SSH_AUTH_SOCK=%s" % self._getSSHSocketpath()
+        out = out.replace("\n\n\n", "\n\n")
+        out = out.replace("\n\n\n", "\n\n")
+        self.writeFile(bashprofile_path, out)
+
+    def _initSSH_ENV(self, force=False):
+        if force or "SSH_AUTH_SOCK" not in os.environ:
+            os.putenv("SSH_AUTH_SOCK", self._getSSHSocketpath())
+            os.environ["SSH_AUTH_SOCK"] = self._getSSHSocketpath()
+
+    def _getSSHSocketpath(self):
+
+        if "SSH_AUTH_SOCK" in os.environ:
+            return(os.environ["SSH_AUTH_SOCK"])
+
+        socketpath = "%s/sshagent_socket" % os.environ.get("HOME", '/root')
+        os.environ['SSH_AUTH_SOCK'] = socketpath
+        return socketpath
+
+    def SSHAgentCheck(self):
+        if "SSH_AUTH_SOCK" not in os.environ:
+            self._initSSH_ENV(True)
+            self._addSSHAgentToBashProfile()
+        if not self.SSHAgentAvailable():
+            self._loadSSHAgent()
+
+    def SSHKeyLoad(self, path, duration=3600 * 24):
+        """
+        @param path is name or full path
+        """
+        self.SSHAgentCheck()
+        if self.SSHAgentCheckKeyIsLoaded(path):
+            return
+        self.logger.info("load ssh key:%s" % path)
+        self.chmod(path, 0o600)
+        cmd = "ssh-add -t %s %s " % (duration, path)
+        self.executeInteractive(cmd)
+
+    def SSHAgentCheckKeyIsLoaded(self, keyNamePath):
+        keysloaded = [self.getBaseName(item)
+                      for item in self.SSHKeysListFromAgent()]
+        if self.getBaseName(keyNamePath) in keysloaded:
+            self.logger.info("ssh key:%s loaded" % keyNamePath)
+            return True
+        else:
+            self.logger.info("ssh key:%s NOT loaded" % keyNamePath)
+            return False
+
+    # def _askItemsFromList(self, items, msg=""):
+    #     if len(items) == 0:
+    #         return []
+    #     if msg != "":
+    #         print(msg)
+    #     nr = 0
+    #     for item in items:
+    #         nr += 1
+    #         print(" - %s: %s" % (nr, item))
+    #     print("select item(s) from list (nr or comma separated list of nr, * is all)")
+    #     item = input()
+    #     if item.strip() == "*":
+    #         return items
+    #     elif item.find(",") != -1:
+    #         res = []
+    #         itemsselected = [item.strip() for item in item.split(",") if item.strip() != ""]
+    #         for item in itemsselected:
+    #             item = int(item) - 1
+    #             res.append(items[item])
+    #         return res
+    #     else:
+    #         item = int(item) - 1
+    #         return [items[item]]
+
+    def SSHKeysLoad(self, path=None, duration=3600 * 24, die=False):
+        """
+        will see if ssh-agent has been started
+        will check keys in home dir
+        will ask which keys to load
+        will adjust .profile file to make sure that env param is set to allow ssh-agent to find the keys
+        """
+        self.SSHAgentCheck()
+
+        if path is None:
+            path = os.path.expanduser("~/.ssh")
+        self.createDir(path)
+
+        if "SSH_AUTH_SOCK" not in os.environ:
+            self._initSSH_ENV(True)
+
+        self._loadSSHAgent()
+
+        keysloaded = [self.getBaseName(item)
+                      for item in self.SSHKeysListFromAgent()]
+
+        if self.isDir(path):
+            keysinfs = [self.getBaseName(item).replace(".pub", "") for item in self.listFilesInDir(
+                path, filter="*.pub") if self.exists(item.replace(".pub", ""))]
+            keysinfs = [item for item in keysinfs if item not in keysloaded]
+
+            res = self.askItemsFromList(
+                keysinfs,
+                "select ssh keys to load, use comma separated list e.g. 1,4,3 and press enter.")
+        else:
+            res = [self.getBaseName(path).replace(".pub", "")]
+            path = self.getParent(path)
+
+        for item in res:
+            pathkey = "%s/%s" % (path, item)
+            # timeout after 24 h
+            self.logger.info("load sshkey: %s" % pathkey)
+            cmd = "ssh-add -t %s %s " % (duration, pathkey)
+            self.executeInteractive(cmd)
+
+    def SSHKeyGetPathFromAgent(self, keyname, die=True):
+        try:
+            # TODO: why do we use subprocess here and not self.execute?
+            out = subprocess.check_output(["ssh-add", "-L"])
+        except BaseException:
+            return None
+
+        for line in out.splitlines():
+            delim = ("/%s" % keyname).encode()
+
+            if line.endswith(delim):
+                line = line.strip()
+                keypath = line.split(" ".encode())[-1]
+                content = line.split(" ".encode())[-2]
+                if not self.exists(path=keypath):
+                    if self.exists("keys/%s" % keyname):
+                        keypath = "keys/%s" % keyname
+                    else:
+                        raise RuntimeError(
+                            "could not find keypath:%s" % keypath)
+                return keypath.decode()
+        if die:
+            raise RuntimeError(
+                "Did not find key with name:%s, check its loaded in ssh-agent with ssh-add -l" %
+                keyname)
+        return None
+
+    def SSHKeyGetFromAgentPub(self, keyname, die=True):
+        try:
+            # TODO: why do we use subprocess here and not self.execute?
+            out = subprocess.check_output(["ssh-add", "-L"])
+        except BaseException:
+            return None
+
+        for line in out.splitlines():
+            delim = (".ssh/%s" % keyname).encode()
+            if line.endswith(delim):
+                content = line.strip()
+                content = content.decode()
+                return content
+        if die:
+            raise RuntimeError(
+                "Did not find key with name:%s, check its loaded in ssh-agent with ssh-add -l" %
+                keyname)
+        return None
+
+    def SSHKeysListFromAgent(self, keyIncluded=False):
+        """
+        returns list of paths
+        """
+        if "SSH_AUTH_SOCK" not in os.environ:
+            self._initSSH_ENV(True)
+        self._loadSSHAgent()
+        cmd = "ssh-add -L"
+        rc, out, err = self.execute(cmd, False, False, die=False)
+        if rc:
+            if rc == 1 and out.find("The agent has no identities") != -1:
+                return []
+            raise RuntimeError("error during listing of keys :%s" % err)
+        keys = [line.split()
+                for line in out.splitlines() if len(line.split()) == 3]
+        if keyIncluded:
+            return list(map(lambda key: key[2:0:-1], keys))
+        else:
+            return list(map(lambda key: key[2], keys))
+
+    def SSHEnsureKeyname(self, keyname="", username="root"):
+        if not self.exists(keyname):
+            rootpath = "/root/.ssh/" if username == "root" else "/home/%s/.ssh/"
+            fullpath = self.joinPaths(rootpath, keyname)
+            if self.exists(fullpath):
+                return fullpath
+        return keyname
+
+    def authorize_user(self, sftp_client, ip_address, keyname, username):
+        basename = self.getBaseName(keyname)
+        tmpfile = "/home/%s/.ssh/%s" % (username, basename)
+        self.logger.info("push key to /home/%s/.ssh/%s" % (username, basename))
+        sftp_client.put(keyname, tmpfile)
+
+        # cannot upload directly to root dir
+        auth_key_path = "/home/%s/.ssh/authorized_keys" % username
+        cmd = "ssh %s@%s 'cat %s | sudo tee -a %s '" % username, ip_address, tmpfile, auth_key_path
+        self.logger.info(
+            "do the following on the console\nsudo -s\ncat %s >> %s" %
+            (tmpfile, auth_key_path))
+        self.logger.info(cmd)
+        self.executeInteractive(cmd)
+
+    def authorize_root(self, sftp_client, ip_address, keyname):
+        tmppath = '/tmp/authorized_keys'
+        auth_key_path = "/root/.ssh/authorized_keys"
+        self.delete(tmppath)
+        try:
+            sftp_client.get(auth_key_path, tmppath)
+        except Exception as e:
+            if str(e).find("No such file") != -1:
+                try:
+                    auth_key_path += "2"
+                    sftp_client.get(auth_key_path, tmppath)
+                except Exception as e:
+                    if str(e).find("No such file") != -1:
+                        self.writeFile(tmppath, "")
+                    else:
+                        raise RuntimeError(
+                            "Could not get authorized key,%s" % e)
+
+            C = self.readFile(tmppath)
+            Cnew = self.readFile(keyname)
+            key = Cnew.split(" ")[1]
+            if C.find(key) == -1:
+                C2 = "%s\n%s\n" % (C.strip(), Cnew)
+                C2 = C2.strip() + "\n"
+                self.writeFile(tmppath, C2)
+                self.logger.info("sshauthorized adjusted")
+                sftp_client.put(tmppath, auth_key_path)
+            else:
+                self.logger.info("ssh key was already authorized")
+
+    def SSHAuthorizeKey(
+            self,
+            remoteipaddr,
+            keyname,
+            login="root",
+            passwd=None,
+            sshport=22,
+            removeothers=False):
+        """
+        this required ssh-agent to be loaded !!!
+        the keyname is the name of the key as loaded in ssh-agent
+
+        if remoteothers==True: then other keys will be removed
+        """
+        keyname = self.SSHEnsureKeyname(keyname=keyname, username=login)
+        import paramiko
+        paramiko.util.log_to_file("/tmp/paramiko.log")
+        ssh = paramiko.SSHClient()
+
+        ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+        self.logger.info("ssh connect:%s %s" % (remoteipaddr, login))
+
+        if not self.SSHKeysListFromAgent(self.getBaseName(keyname)):
+            self.SSHKeysLoad(self.getParent(keyname))
+        ssh.connect(
+            remoteipaddr,
+            username=login,
+            password=passwd,
+            allow_agent=True,
+            look_for_keys=False)
+        self.logger.info("ok")
+
+        ftp = ssh.open_sftp()
+
+        if login != "root":
+            self.authorize_user(
+                sftp_client=ftp,
+                ip_address=remoteipaddr,
+                keyname=keyname,
+                username=login)
+        else:
+            self.authorize_root(
+                sftp_client=ftp,
+                ip_address=remoteipaddr,
+                keyname=keyname)
+
+    def _loadSSHAgent(self, path=None, createkeys=False, killfirst=False):
+        """
+        check if ssh-agent is available & there is key loaded
+
+        @param path: is path to private ssh key
+
+        the primary key is 'id_rsa' and will be used as default e.g. if authorizing another node then this key will be used
+
+        """
+        # check if more than 1 agent
+        socketpath = self._getSSHSocketpath()
+        res = [
+            item for item in self.execute(
+                "ps aux|grep ssh-agent",
+                False,
+                False)[1].split("\n") if item.find("grep ssh-agent") == -
+            1]
+        res = [item for item in res if item.strip() != ""]
+        res = [item for item in res if item[-2:] != "-l"]
+
+        if len(res) > 1:
+            self.logger.info("more than 1 ssh-agent, will kill all")
+            killfirst = True
+        if len(res) == 0 and self.exists(socketpath):
+            self.delete(socketpath)
+
+        if killfirst:
+            cmd = "killall ssh-agent"
+            # self.logger.info(cmd)
+            self.execute(cmd, showout=False, outputStderr=False, die=False)
+            # remove previous socketpath
+            self.delete(socketpath)
+            self.delete(self.joinPaths('/tmp', "ssh-agent-pid"))
+
+        if not self.exists(socketpath):
+            self.createDir(self.getParent(socketpath))
+            # ssh-agent not loaded
+            self.logger.info("load ssh agent")
+            rc, result, err = self.execute(
+                "ssh-agent -a %s" %
+                socketpath, die=False, showout=False, outputStderr=False)
+
+            if rc > 0:
+                # could not start ssh-agent
+                raise RuntimeError(
+                    "Could not start ssh-agent, something went wrong,\nstdout:%s\nstderr:%s\n" %
+                    (result, err))
+            else:
+                # get pid from result of ssh-agent being started
+                if not self.exists(socketpath):
+                    raise RuntimeError(
+                        "Serious bug, ssh-agent not started while there was no error, should never get here")
+                piditems = [item for item in result.split(
+                    "\n") if item.find("pid") != -1]
+                # print(piditems)
+                if len(piditems) < 1:
+                    print("results was:")
+                    print(result)
+                    print("END")
+                    raise RuntimeError("Cannot find items in ssh-add -l")
+                self._initSSH_ENV(True)
+                pid = int(piditems[-1].split(" ")[-1].strip("; "))
+                self.writeFile(
+                    self.joinPaths(
+                        '/tmp',
+                        "ssh-agent-pid"),
+                    str(pid))
+                self._addSSHAgentToBashProfile()
+
+            # ssh agent should be loaded because ssh-agent socket has been
+            # found
+            if os.environ.get("SSH_AUTH_SOCK") != socketpath:
+                self._initSSH_ENV(True)
+            rc, result, err = self.execute(
+                "ssh-add -l", die=False, showout=False, outputStderr=False)
+            if rc == 2:
+                # no ssh-agent found
+                print(result)
+                raise RuntimeError(
+                    "Could not connect to ssh-agent, this is bug, ssh-agent should be loaded by now")
+            elif rc == 1:
+                # no keys but agent loaded
+                result = ""
+            elif rc > 0:
+                raise RuntimeError(
+                    "Could not start ssh-agent, something went wrong,\nstdout:%s\nstderr:%s\n" %
+                    (result, err))
+
+    def SSHAgentAvailable(self):
+        if not self.exists(self._getSSHSocketpath()):
+            return False
+        if "SSH_AUTH_SOCK" not in os.environ:
+            self._initSSH_ENV(True)
+        rc, out, err = self.execute(
+            "ssh-add -l", showout=False, outputStderr=False, die=False)
+        if 'The agent has no identities.' in out:
+            return True
+        if rc != 0:
+            return False
+        else:
+            return True
+
+
