+++ JumpScale9/core/InstallTools.py
@@ -30,765 +30,6 @@ class TimeoutError(RuntimeError, TimeoutError):
     pass
 
 
-class SSHMethods():
-
-    def _addSSHAgentToBashProfile(self, path=None):
-
-        bashprofile_path = os.path.expanduser("~/.bash_profile")
-        if not self.exists(bashprofile_path):
-            self.execute('touch %s' % bashprofile_path)
-
-        content = self.readFile(bashprofile_path)
-        out = ""
-        for line in content.split("\n"):
-            if line.find("#JSSSHAGENT") != -1:
-                continue
-            if line.find("SSH_AUTH_SOCK") != -1:
-                continue
-
-            out += "%s\n" % line
-
-        if "SSH_AUTH_SOCK" in os.environ:
-            self.logger.info("NO NEED TO ADD SSH_AUTH_SOCK to env")
-            self.writeFile(bashprofile_path, out)
-            return
-
-        # out += "\njs 'j.do._.loadSSHAgent()' #JSSSHAGENT\n"
-        out += "export SSH_AUTH_SOCK=%s" % self._getSSHSocketpath()
-        out = out.replace("\n\n\n", "\n\n")
-        out = out.replace("\n\n\n", "\n\n")
-        self.writeFile(bashprofile_path, out)
-
-    def _initSSH_ENV(self, force=False):
-        if force or "SSH_AUTH_SOCK" not in os.environ:
-            os.putenv("SSH_AUTH_SOCK", self._getSSHSocketpath())
-            os.environ["SSH_AUTH_SOCK"] = self._getSSHSocketpath()
-
-    def _getSSHSocketpath(self):
-
-        if "SSH_AUTH_SOCK" in os.environ:
-            return(os.environ["SSH_AUTH_SOCK"])
-
-        socketpath = "%s/sshagent_socket" % os.environ.get("HOME", '/root')
-        os.environ['SSH_AUTH_SOCK'] = socketpath
-        return socketpath
-
-    def SSHAgentCheck(self):
-        if "SSH_AUTH_SOCK" not in os.environ:
-            self._initSSH_ENV(True)
-            self._addSSHAgentToBashProfile()
-        if not self.SSHAgentAvailable():
-            self._loadSSHAgent()
-
-    def SSHKeyLoad(self, path, duration=3600 * 24):
-        """
-        @param path is name or full path
-        """
-        self.SSHAgentCheck()
-        if self.SSHAgentCheckKeyIsLoaded(path):
-            return
-        self.logger.info("load ssh key:%s" % path)
-        self.chmod(path, 0o600)
-        cmd = "ssh-add -t %s %s " % (duration, path)
-        self.executeInteractive(cmd)
-
-    def SSHAgentCheckKeyIsLoaded(self, keyNamePath):
-        keysloaded = [self.getBaseName(item)
-                      for item in self.SSHKeysListFromAgent()]
-        if self.getBaseName(keyNamePath) in keysloaded:
-            self.logger.info("ssh key:%s loaded" % keyNamePath)
-            return True
-        else:
-            self.logger.info("ssh key:%s NOT loaded" % keyNamePath)
-            return False
-
-    def askItemsFromList(self, items, msg=""):
-        if len(items) == 0:
-            return []
-        if msg != "":
-            print(msg)
-        nr = 0
-        for item in items:
-            nr += 1
-            print(" - %s: %s" % (nr, item))
-        print("select item(s) from list (nr or comma separated list of nr, * is all)")
-        item = input()
-        if item.strip() == "*":
-            return items
-        elif item.find(",") != -1:
-            res = []
-            itemsselected = [item.strip() for item in item.split(",") if item.strip() != ""]
-            for item in itemsselected:
-                item = int(item) - 1
-                res.append(items[item])
-            return res
-        else:
-            item = int(item) - 1
-            return [items[item]]
-
-    def SSHKeysLoad(self, path=None, duration=3600 * 24, die=False):
-        """
-        will see if ssh-agent has been started
-        will check keys in home dir
-        will ask which keys to load
-        will adjust .profile file to make sure that env param is set to allow ssh-agent to find the keys
-        """
-        self.SSHAgentCheck()
-
-        if path is None:
-            path = os.path.expanduser("~/.ssh")
-        self.createDir(path)
-
-        if "SSH_AUTH_SOCK" not in os.environ:
-            self._initSSH_ENV(True)
-
-        self._loadSSHAgent()
-
-        keysloaded = [self.getBaseName(item)
-                      for item in self.SSHKeysListFromAgent()]
-
-        if self.isDir(path):
-            keysinfs = [self.getBaseName(item).replace(".pub", "") for item in self.listFilesInDir(
-                path, filter="*.pub") if self.exists(item.replace(".pub", ""))]
-            keysinfs = [item for item in keysinfs if item not in keysloaded]
-
-            res = self.askItemsFromList(
-                keysinfs,
-                "select ssh keys to load, use comma separated list e.g. 1,4,3 and press enter.")
-        else:
-            res = [self.getBaseName(path).replace(".pub", "")]
-            path = self.getParent(path)
-
-        for item in res:
-            pathkey = "%s/%s" % (path, item)
-            # timeout after 24 h
-            self.logger.info("load sshkey: %s" % pathkey)
-            cmd = "ssh-add -t %s %s " % (duration, pathkey)
-            self.executeInteractive(cmd)
-
-    def SSHKeyGetPathFromAgent(self, keyname, die=True):
-        try:
-            # TODO: why do we use subprocess here and not self.execute?
-            out = subprocess.check_output(["ssh-add", "-L"])
-        except BaseException:
-            return None
-
-        for line in out.splitlines():
-            delim = ("/%s" % keyname).encode()
-
-            if line.endswith(delim):
-                line = line.strip()
-                keypath = line.split(" ".encode())[-1]
-                content = line.split(" ".encode())[-2]
-                if not self.exists(path=keypath):
-                    if self.exists("keys/%s" % keyname):
-                        keypath = "keys/%s" % keyname
-                    else:
-                        raise RuntimeError(
-                            "could not find keypath:%s" % keypath)
-                return keypath.decode()
-        if die:
-            raise RuntimeError(
-                "Did not find key with name:%s, check its loaded in ssh-agent with ssh-add -l" %
-                keyname)
-        return None
-
-    def SSHKeyGetFromAgentPub(self, keyname, die=True):
-        try:
-            # TODO: why do we use subprocess here and not self.execute?
-            out = subprocess.check_output(["ssh-add", "-L"])
-        except BaseException:
-            return None
-
-        for line in out.splitlines():
-            delim = (".ssh/%s" % keyname).encode()
-            if line.endswith(delim):
-                content = line.strip()
-                content = content.decode()
-                return content
-        if die:
-            raise RuntimeError(
-                "Did not find key with name:%s, check its loaded in ssh-agent with ssh-add -l" %
-                keyname)
-        return None
-
-    def SSHKeysListFromAgent(self, keyIncluded=False):
-        """
-        returns list of paths
-        """
-        if "SSH_AUTH_SOCK" not in os.environ:
-            self._initSSH_ENV(True)
-        self._loadSSHAgent()
-        cmd = "ssh-add -L"
-        rc, out, err = self.execute(cmd, False, False, die=False)
-        if rc:
-            if rc == 1 and out.find("The agent has no identities") != -1:
-                return []
-            raise RuntimeError("error during listing of keys :%s" % err)
-        keys = [line.split()
-                for line in out.splitlines() if len(line.split()) == 3]
-        if keyIncluded:
-            return list(map(lambda key: key[2:0:-1], keys))
-        else:
-            return list(map(lambda key: key[2], keys))
-
-    def SSHEnsureKeyname(self, keyname="", username="root"):
-        if not self.exists(keyname):
-            rootpath = "/root/.ssh/" if username == "root" else "/home/%s/.ssh/"
-            fullpath = self.joinPaths(rootpath, keyname)
-            if self.exists(fullpath):
-                return fullpath
-        return keyname
-
-    def authorize_user(self, sftp_client, ip_address, keyname, username):
-        basename = self.getBaseName(keyname)
-        tmpfile = "/home/%s/.ssh/%s" % (username, basename)
-        self.logger.info("push key to /home/%s/.ssh/%s" % (username, basename))
-        sftp_client.put(keyname, tmpfile)
-
-        # cannot upload directly to root dir
-        auth_key_path = "/home/%s/.ssh/authorized_keys" % username
-        cmd = "ssh %s@%s 'cat %s | sudo tee -a %s '" % username, ip_address, tmpfile, auth_key_path
-        self.logger.info(
-            "do the following on the console\nsudo -s\ncat %s >> %s" %
-            (tmpfile, auth_key_path))
-        self.logger.info(cmd)
-        self.executeInteractive(cmd)
-
-    def authorize_root(self, sftp_client, ip_address, keyname):
-        tmppath = '/tmp/authorized_keys'
-        auth_key_path = "/root/.ssh/authorized_keys"
-        self.delete(tmppath)
-        try:
-            sftp_client.get(auth_key_path, tmppath)
-        except Exception as e:
-            if str(e).find("No such file") != -1:
-                try:
-                    auth_key_path += "2"
-                    sftp_client.get(auth_key_path, tmppath)
-                except Exception as e:
-                    if str(e).find("No such file") != -1:
-                        self.writeFile(tmppath, "")
-                    else:
-                        raise RuntimeError(
-                            "Could not get authorized key,%s" % e)
-
-            C = self.readFile(tmppath)
-            Cnew = self.readFile(keyname)
-            key = Cnew.split(" ")[1]
-            if C.find(key) == -1:
-                C2 = "%s\n%s\n" % (C.strip(), Cnew)
-                C2 = C2.strip() + "\n"
-                self.writeFile(tmppath, C2)
-                self.logger.info("sshauthorized adjusted")
-                sftp_client.put(tmppath, auth_key_path)
-            else:
-                self.logger.info("ssh key was already authorized")
-
-    def SSHAuthorizeKey(
-            self,
-            remoteipaddr,
-            keyname,
-            login="root",
-            passwd=None,
-            sshport=22,
-            removeothers=False):
-        """
-        this required ssh-agent to be loaded !!!
-        the keyname is the name of the key as loaded in ssh-agent
-
-        if remoteothers==True: then other keys will be removed
-        """
-        keyname = self.SSHEnsureKeyname(keyname=keyname, username=login)
-        import paramiko
-        paramiko.util.log_to_file("/tmp/paramiko.log")
-        ssh = paramiko.SSHClient()
-
-        ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        self.logger.info("ssh connect:%s %s" % (remoteipaddr, login))
-
-        if not self.SSHKeysListFromAgent(self.getBaseName(keyname)):
-            self.SSHKeysLoad(self.getParent(keyname))
-        ssh.connect(
-            remoteipaddr,
-            username=login,
-            password=passwd,
-            allow_agent=True,
-            look_for_keys=False)
-        self.logger.info("ok")
-
-        ftp = ssh.open_sftp()
-
-        if login != "root":
-            self.authorize_user(
-                sftp_client=ftp,
-                ip_address=remoteipaddr,
-                keyname=keyname,
-                username=login)
-        else:
-            self.authorize_root(
-                sftp_client=ftp,
-                ip_address=remoteipaddr,
-                keyname=keyname)
-
-    def _loadSSHAgent(self, path=None, createkeys=False, killfirst=False):
-        """
-        check if ssh-agent is available & there is key loaded
-
-        @param path: is path to private ssh key
-
-        the primary key is 'id_rsa' and will be used as default e.g. if authorizing another node then this key will be used
-
-        """
-        # check if more than 1 agent
-        socketpath = self._getSSHSocketpath()
-        res = [
-            item for item in self.execute(
-                "ps aux|grep ssh-agent",
-                False,
-                False)[1].split("\n") if item.find("grep ssh-agent") == -
-            1]
-        res = [item for item in res if item.strip() != ""]
-        res = [item for item in res if item[-2:] != "-l"]
-
-        if len(res) > 1:
-            self.logger.info("more than 1 ssh-agent, will kill all")
-            killfirst = True
-        if len(res) == 0 and self.exists(socketpath):
-            self.delete(socketpath)
-
-        if killfirst:
-            cmd = "killall ssh-agent"
-            # self.logger.info(cmd)
-            self.execute(cmd, showout=False, outputStderr=False, die=False)
-            # remove previous socketpath
-            self.delete(socketpath)
-            self.delete(self.joinPaths('/tmp', "ssh-agent-pid"))
-
-        if not self.exists(socketpath):
-            self.createDir(self.getParent(socketpath))
-            # ssh-agent not loaded
-            self.logger.info("load ssh agent")
-            rc, result, err = self.execute(
-                "ssh-agent -a %s" %
-                socketpath, die=False, showout=False, outputStderr=False)
-
-            if rc > 0:
-                # could not start ssh-agent
-                raise RuntimeError(
-                    "Could not start ssh-agent, something went wrong,\nstdout:%s\nstderr:%s\n" %
-                    (result, err))
-            else:
-                # get pid from result of ssh-agent being started
-                if not self.exists(socketpath):
-                    raise RuntimeError(
-                        "Serious bug, ssh-agent not started while there was no error, should never get here")
-                piditems = [item for item in result.split(
-                    "\n") if item.find("pid") != -1]
-                # print(piditems)
-                if len(piditems) < 1:
-                    print("results was:")
-                    print(result)
-                    print("END")
-                    raise RuntimeError("Cannot find items in ssh-add -l")
-                self._initSSH_ENV(True)
-                pid = int(piditems[-1].split(" ")[-1].strip("; "))
-                self.writeFile(
-                    self.joinPaths(
-                        '/tmp',
-                        "ssh-agent-pid"),
-                    str(pid))
-                self._addSSHAgentToBashProfile()
-
-            # ssh agent should be loaded because ssh-agent socket has been
-            # found
-            if os.environ.get("SSH_AUTH_SOCK") != socketpath:
-                self._initSSH_ENV(True)
-            rc, result, err = self.execute(
-                "ssh-add -l", die=False, showout=False, outputStderr=False)
-            if rc == 2:
-                # no ssh-agent found
-                print(result)
-                raise RuntimeError(
-                    "Could not connect to ssh-agent, this is bug, ssh-agent should be loaded by now")
-            elif rc == 1:
-                # no keys but agent loaded
-                result = ""
-            elif rc > 0:
-                raise RuntimeError(
-                    "Could not start ssh-agent, something went wrong,\nstdout:%s\nstderr:%s\n" %
-                    (result, err))
-
-    def SSHAgentAvailable(self):
-        if not self.exists(self._getSSHSocketpath()):
-            return False
-        if "SSH_AUTH_SOCK" not in os.environ:
-            self._initSSH_ENV(True)
-        rc, out, err = self.execute(
-            "ssh-add -l", showout=False, outputStderr=False, die=False)
-        if 'The agent has no identities.' in out:
-            return True
-        if rc != 0:
-            return False
-        else:
-            return True
-
-
-class GitMethods():
-
-    def rewriteGitRepoUrl(self, url="", login=None, passwd=None, ssh="auto"):
-        """
-        Rewrite the url of a git repo with login and passwd if specified
-
-        Args:
-            url (str): the HTTP URL of the Git repository. ex: 'https://github.com/odoo/odoo'
-            login (str): authentication login name
-            passwd (str): authentication login password
-            ssh = if True will build ssh url, if "auto" will check if there is ssh-agent available & keys are loaded,
-                if yes will use ssh
-
-        Returns:
-            (repository_host, repository_type, repository_account, repository_name, repository_url)
-        """
-
-        if ssh == "auto" or ssh == "first":
-            ssh = self.SSHAgentAvailable()
-        elif ssh or ssh is False:
-            pass
-        else:
-            raise RuntimeError(
-                "ssh needs to be auto, first or True or False: here:'%s'" %
-                ssh)
-
-        url_pattern_ssh = re.compile('^(git@)(.*?):(.*?)/(.*?)/?$')
-        sshmatch = url_pattern_ssh.match(url)
-        url_pattern_http = re.compile('^(https?://)(.*?)/(.*?)/(.*?)/?$')
-        httpmatch = url_pattern_http.match(url)
-        if not sshmatch:
-            match = httpmatch
-        else:
-            match = sshmatch
-
-        if not match:
-            raise RuntimeError(
-                "Url is invalid. Must be in the form of 'http(s)://hostname/account/repo' or 'git@hostname:account/repo'")
-
-        protocol, repository_host, repository_account, repository_name = match.groups()
-        if protocol.startswith("git") and ssh is False:
-            protocol = "https://"
-
-        if not repository_name.endswith('.git'):
-            repository_name += '.git'
-
-        if login == 'ssh' or ssh:
-            repository_url = 'git@%(host)s:%(account)s/%(name)s' % {
-                'host': repository_host,
-                'account': repository_account,
-                'name': repository_name,
-            }
-            protocol = "ssh"
-
-        elif login and login != 'guest':
-            repository_url = '%(protocol)s%(login)s:%(password)s@%(host)s/%(account)s/%(repo)s' % {
-                'protocol': protocol,
-                'login': login,
-                'password': passwd,
-                'host': repository_host,
-                'account': repository_account,
-                'repo': repository_name,
-            }
-
-        else:
-            repository_url = '%(protocol)s%(host)s/%(account)s/%(repo)s' % {
-                'protocol': protocol,
-                'host': repository_host,
-                'account': repository_account,
-                'repo': repository_name,
-            }
-        if repository_name.endswith(".git"):
-            repository_name = repository_name[:-4]
-
-        return protocol, repository_host, repository_account, repository_name, repository_url
-
-    def getGitRepoArgs(
-            self,
-            url="",
-            dest=None,
-            login=None,
-            passwd=None,
-            reset=False,
-            branch=None,
-            ssh="auto",
-            codeDir=None,
-            executor=None):
-        """
-        Extracts and returns data useful in cloning a Git repository.
-
-        Args:
-            url (str): the HTTP/GIT URL of the Git repository to clone from. eg: 'https://github.com/odoo/odoo.git'
-            dest (str): the local filesystem path to clone to
-            login (str): authentication login name (only for http)
-            passwd (str): authentication login password (only for http)
-            reset (boolean): if True, any cached clone of the Git repository will be removed
-            branch (str): branch to be used
-            ssh if auto will check if ssh-agent loaded, if True will be forced to use ssh for git
-
-        # Process for finding authentication credentials (NOT IMPLEMENTED YET)
-
-        - first check there is an ssh-agent and there is a key attached to it, if yes then no login & passwd will be used & method will always be git
-        - if not ssh-agent found
-            - then we will check if url is github & ENV argument GITHUBUSER & GITHUBPASSWD is set
-                - if env arguments set, we will use those & ignore login/passwd arguments
-            - we will check if login/passwd specified in URL, if yes willl use those (so they get priority on login/passwd arguments)
-            - we will see if login/passwd specified as arguments, if yes will use those
-        - if we don't know login or passwd yet then
-            - login/passwd will be fetched from local git repo directory (if it exists and reset==False)
-        - if at this point still no login/passwd then we will try to build url with anonymous
-
-
-        # Process for defining branch
-
-        - if branch arg: None
-            - check if git directory exists if yes take that branch
-            - default to 'master'
-        - if it exists, use the branch arg
-
-        Returns:
-            (repository_host, repository_type, repository_account, repository_name, dest, repository_url)
-
-            - repository_type http or git
-
-        Remark:
-            url can be empty, then the git params will be fetched out of the git configuration at that path
-        """
-
-        if url == "":
-            if dest is None:
-                raise RuntimeError("dest cannot be None (url is also '')")
-            if not self.exists(dest):
-                raise RuntimeError(
-                    "Could not find git repo path:%s, url was not specified so git destination needs to be specified." %
-                    (dest))
-
-        if login is None and url.find("github.com/") != -1:
-            # can see if there if login & passwd in OS env
-            # if yes fill it in
-            if "GITHUBUSER" in os.environ:
-                login = os.environ["GITHUBUSER"]
-            if "GITHUBPASSWD" in os.environ:
-                passwd = os.environ["GITHUBPASSWD"]
-
-        protocol, repository_host, repository_account, repository_name, repository_url = self.rewriteGitRepoUrl(
-            url=url, login=login, passwd=passwd, ssh=ssh)
-
-        repository_type = repository_host.split(
-            '.')[0] if '.' in repository_host else repository_host
-
-        if not dest:
-            if codeDir is None:
-                if not executor:
-                    codeDir = j.CODEDIR
-                else:
-                    codeDir = executor.prefab.core.dir_paths['CODEDIR']
-            dest = '%(codedir)s/%(type)s/%(account)s/%(repo_name)s' % {
-                'codedir': codeDir,
-                'type': repository_type.lower(),
-                'account': repository_account.lower(),
-                'repo_name': repository_name,
-            }
-
-        if reset:
-            self.delete(dest)
-
-        # self.createDir(dest)
-
-        return repository_host, repository_type, repository_account, repository_name, dest, repository_url
-
-    def pullGitRepo(
-            self,
-            url="",
-            dest=None,
-            login=None,
-            passwd=None,
-            depth=None,
-            ignorelocalchanges=False,
-            reset=False,
-            branch=None,
-            tag=None,
-            revision=None,
-            ssh="auto",
-            executor=None,
-            codeDir=None,
-            timeout=600):
-        """
-        will clone or update repo
-        if dest is None then clone underneath: /opt/code/$type/$account/$repo
-        will ignore changes !!!!!!!!!!!
-
-        @param ssh ==True means will checkout ssh
-        @param ssh =="first" means will checkout sss first if that does not work will go to http
-        """
-        if branch == "":
-            branch = None
-        if branch is not None and tag is not None:
-            raise RuntimeError("only branch or tag can be set")
-
-        if ssh == "first" or ssh == "auto":
-            try:
-                return self.pullGitRepo(
-                    url,
-                    dest,
-                    login,
-                    passwd,
-                    depth,
-                    ignorelocalchanges,
-                    reset,
-                    branch,
-                    tag=tag,
-                    revision=revision,
-                    ssh=True,
-                    executor=executor)
-            except Exception as e:
-                base, provider, account, repo, dest, url = self.getGitRepoArgs(
-                    url, dest, login, passwd, reset=reset, ssh=False, codeDir=codeDir, executor=executor)
-                return self.pullGitRepo(
-                    url,
-                    dest,
-                    login,
-                    passwd,
-                    depth,
-                    ignorelocalchanges,
-                    reset,
-                    branch,
-                    tag=tag,
-                    revision=revision,
-                    ssh=False,
-                    executor=executor)
-
-        base, provider, account, repo, dest, url = self.getGitRepoArgs(
-            url, dest, login, passwd, reset=reset, ssh=ssh, codeDir=codeDir, executor=executor)
-
-        self.logger.info("%s:pull:%s ->%s" % (executor, url, dest))
-
-        existsDir = self.exists(
-            dest) if not executor else executor.exists(dest)
-
-        checkdir = "%s/.git" % (dest)
-        existsGit = self.exists(
-            checkdir) if not executor else executor.exists(checkdir)
-
-        if existsDir:
-            if not existsGit:
-                raise RuntimeError("found directory but .git not found in %s" % dest)
-            # if we don't specify the branch, try to find the currently
-            # checkedout branch
-            cmd = 'cd %s; git rev-parse --abbrev-ref HEAD' % dest
-            rc, out, err = self.execute(
-                cmd, die=False, showout=False, executor=executor)
-            if rc == 0:
-                branchFound = out.strip()
-            else:  # if we can't retreive current branch, use master as default
-                branchFound = 'master'
-                # raise RuntimeError("Cannot retrieve branch:\n%s\n" % cmd)
-            if branch is not None and branch != branchFound and ignorelocalchanges is False:
-                raise RuntimeError(
-                    "Cannot pull repo '%s', branch on filesystem is not same as branch asked for.\n"
-                    "Branch asked for: %s\n"
-                    "Branch found: %s\n"
-                    "To choose other branch do e.g:"
-                    "export JSBRANCH='%s'\n" %
-                    (repo, branch, branchFound, branchFound))
-
-            if ignorelocalchanges:
-                self.logger.info(
-                    ("git pull, ignore changes %s -> %s" %
-                     (url, dest)))
-                cmd = "cd %s;git fetch" % dest
-                if depth is not None:
-                    cmd += " --depth %s" % depth
-                    self.execute(cmd, executor=executor)
-                if branch is not None:
-                    self.logger.info("reset branch to:%s" % branch)
-                    self.execute(
-                        "cd %s;git fetch; git reset --hard origin/%s" %
-                        (dest, branch), timeout=timeout, executor=executor)
-
-            else:
-
-                if branch is None and tag is None:
-                    branch = branchFound
-
-                # pull
-                self.logger.info(("git pull %s -> %s" % (url, dest)))
-                if url.find("http") != -1:
-                    cmd = "mkdir -p %s;cd %s;git -c http.sslVerify=false pull origin %s" % (
-                        dest, dest, branch)
-                    self.logger.info(cmd)
-                    self.execute(cmd, timeout=timeout, executor=executor)
-                else:
-                    try:
-                        cmd = "cd %s;git pull origin %s" % (dest, branch)
-                        self.logger.info(cmd)
-                        self.execute(cmd, timeout=timeout, executor=executor)
-                    except Exception as e:
-                        protocol, host, account, repo_name, repo_url = self.rewriteGitRepoUrl(url=url, ssh=False)
-                        cmd = "cd %s;git -c http.sslVerify=false pull %s %s" % (dest, repo_url, branch)
-                        self.logger.info(cmd)
-                        self.execute(cmd, timeout=timeout, executor=executor)
-
-        else:
-            self.logger.info(("git clone %s -> %s" % (url, dest)))
-            # self.createDir(dest)
-            extra = ""
-            if depth is not None:
-                extra = "--depth=%s" % depth
-            if url.find("http") != -1:
-                if branch is not None:
-                    cmd = "mkdir -p %s;cd %s;git -c http.sslVerify=false clone %s -b %s %s %s" % (
-                        self.getParent(dest), self.getParent(dest), extra, branch, url, dest)
-                else:
-                    cmd = "mkdir -p %s;cd %s;git -c http.sslVerify=false clone %s  %s %s" % (
-                        self.getParent(dest), self.getParent(dest), extra, url, dest)
-            else:
-                if branch is not None:
-                    cmd = "mkdir -p %s;cd %s;git clone %s -b %s %s %s" % (
-                        self.getParent(dest), self.getParent(dest), extra, branch, url, dest)
-                else:
-                    cmd = "mkdir -p %s;cd %s;git clone %s  %s %s" % (
-                        self.getParent(dest), self.getParent(dest), extra, url, dest)
-
-            self.logger.info(cmd)
-
-            # self.logger.info(str(executor)+" "+cmd)
-            self.execute(cmd, timeout=timeout, executor=executor)
-
-        if tag is not None:
-            self.logger.info("reset tag to:%s" % tag)
-            self.execute("cd %s;git checkout tags/%s" %
-                         (dest, tag), timeout=timeout, executor=executor)
-
-        if revision is not None:
-            cmd = "mkdir -p %s;cd %s;git checkout %s" % (dest, dest, revision)
-            self.logger.info(cmd)
-            self.execute(cmd, timeout=timeout, executor=executor)
-
-        return dest
-
-    def getGitBranch(self, path):
-
-        # if we don't specify the branch, try to find the currently checkedout
-        # branch
-        cmd = 'cd %s;git rev-parse --abbrev-ref HEAD' % path
-        try:
-            rc, out, err = self.execute(cmd, showout=False, outputStderr=False)
-            if rc == 0:
-                branch = out.strip()
-            else:  # if we can't retreive current branch, use master as default
-                branch = 'master'
-        except BaseException:
-            branch = 'master'
-
-        return branch
 
 
 class FSMethods():
@@ -1763,9 +1004,9 @@ class FSMethods():
         return "%s/jumpscaleinstall/%s" % ('/tmp', filename)
 
 
-    def getWalker(self):
-        self._initExtra()
-        return self.extra.getWalker(self)
+    # def getWalker(self):
+    #     self._initExtra()
+    #     return self.extra.getWalker(self)
 
 
 class ExecutorMethods():
@@ -1927,8 +1168,7 @@ class ExecutorMethods():
                 name)
 
 
-class InstallTools(GitMethods, FSMethods, ExecutorMethods, SSHMethods):
-
+class InstallTools(ExecutorMethods, SSHMethods):
     def __init__(self, debug=False):
 
         self.__jslocation__ = "j.core.installtools"
